+++ src/tuple/index.ts
@@ -1,0 +1,6 @@
+import { type Either } from "./Either";
+import { mightFail, mightFailSync } from "./mightFail";
+import { makeMightFail, makeMightFailSync } from "./makeMightFail";
+
+export { Either, mightFail, makeMightFail, mightFailSync, makeMightFailSync };
+export default { mightFail, makeMightFail, mightFailSync, makeMightFailSync };